// src/types/SummaryMode.ts
export enum SummaryMode {
  PRIOR_TO_TRACKING_MODE = "PRIOR_TO_TRACKING_MODE",
  OVERALL = "OVERALL",
  PER_PAGE = "PER_PAGE",
  EACH_FILE_OVERALL = "EACH_FILE_OVERALL",
  EACH_FILE_IN_CHUNKS = "EACH_FILE_IN_CHUNKS",
  EACH_FILE_PER_PAGE = "EACH_FILE_PER_PAGE",
  EACH_FILE_PER_PAGE_THEN_OVERALL = "EACH_FILE_PER_PAGE_THEN_OVERALL",
  EACH_FILE_OVERLAPPING_CHUNKS = "EACH_FILE_OVERLAPPING_CHUNKS",
}